Reconsider method return values (#35)
- Change types of some internal fields and return values to usize
- Change some return values from `&Option<T>` to `Option<&T>`

+++ src/entities/activity.rs
@@ -1,41 +1,41 @@
 use serde::Deserialize;
-use crate::utils::transform_string_to_u64;
+use crate::utils::transform_string_to_usize;
 use super::Entity;
 
 /// Represents a weekly bucket of instance activity.
 #[derive(Debug, PartialEq, PartialOrd, Hash, Clone, Copy, Deserialize)]
 pub struct Activity {
-    #[serde(deserialize_with = "transform_string_to_u64")]
-    week: u64,
+    #[serde(deserialize_with = "transform_string_to_usize")]
+    week: usize,
 
-    #[serde(deserialize_with = "transform_string_to_u64")]
-    statuses: u64,
+    #[serde(deserialize_with = "transform_string_to_usize")]
+    statuses: usize,
 
-    #[serde(deserialize_with = "transform_string_to_u64")]
-    logins: u64,
+    #[serde(deserialize_with = "transform_string_to_usize")]
+    logins: usize,
 
-    #[serde(deserialize_with = "transform_string_to_u64")]
-    registrations: u64,
+    #[serde(deserialize_with = "transform_string_to_usize")]
+    registrations: usize,
 }
 
 impl Activity {
     /// Get midnight at the first day of the week.
-    pub fn week(&self) -> u64 {
+    pub fn week(&self) -> usize {
         self.week
     }
 
     /// Get statuses created since the week began.
-    pub fn statuses(&self) -> u64 {
+    pub fn statuses(&self) -> usize {
         self.statuses
     }
 
     /// Get user logins since the week began.
-    pub fn logins(&self) -> u64 {
+    pub fn logins(&self) -> usize {
         self.logins
     }
 
     /// Get user registrations since the week began.
-    pub fn registrations(&self) -> u64 {
+    pub fn registrations(&self) -> usize {
         self.registrations
     }
 }
